Separando pygames em arquivo e implementação do randomized growth
