+++ src/generator.py
@@ -1,18 +1,10 @@
 from maze import Maze
+from graphics import *
 import pygame , time, random
 
-WIDTH = 1000
-HEIGHT = 1000
 FPS = 30
 
-w = 10
-w2 = 50
-
-WHITE = (255, 255, 255)
-GREEN = (0, 255, 0,)
-BLUE = (0, 0, 255)
-YELLOW = (255 ,255 ,0)
-screen = pygame.display.set_mode((WIDTH, HEIGHT))
+randomizeStack = True
 
 def DFSmaze(maze, startPoint):
     x_c, y_c = startPoint
@@ -23,6 +15,8 @@ def DFSmaze(maze, startPoint):
     stack.append((x_c, y_c))
     path.append((x_c, y_c))
     while stack:
+        if randomizeStack:
+            random.shuffle(stack)
         x_c, y_c = stack.pop()
         neighbours = maze.findNeighbours(x_c, y_c)
 
@@ -37,56 +31,13 @@ def DFSmaze(maze, startPoint):
                     stack.append(n)
     return path
 
-def gameUp(y, x):
-    x = (x*2 + 1)*w
-    y = (y*2 + 1)*w
-    pygame.draw.rect(screen, WHITE, (x, y-w, w, w*2), 0)
-  
-def gameDown(y, x):
-    x = (x*2 + 1)*w 
-    y = (y*2 + 1)*w
-    pygame.draw.rect(screen, WHITE, (x, y, w, w*2), 0)
-
-def gameLeft(y, x):
-    x = (x*2 + 1)*w
-    y = (y*2 + 1)*w
-    pygame.draw.rect(screen, WHITE, (x-w, y, w*2, w), 0)
-
-def gameRight(y, x):
-    x = (x*2 + 1)*w
-    y = (y*2 + 1)*w
-    pygame.draw.rect(screen, WHITE, (x, y, w*2, w), 0)
-   
-def gameCurrentNode(node):
-    walls = node.walls
-
-    if not walls["down"]:
-        gameDown(node.x,node.y)
-    if not walls["up"]:
-        gameUp(node.x,node.y)
-    if not walls["left"]:
-        gameLeft(node.x,node.y)
-    if not walls["right"]:
-        gameRight(node.x,node.y)
-
-    pygame.display.update()
-    time.sleep(.005)
-
-pygame.init()
-pygame.mixer.init()
-pygame.display.set_caption("Finder")
-clock = pygame.time.Clock()
-
+clock = initGame()
 maze = Maze(50, 30)
 path = DFSmaze(maze, (0, 0))
 
 for node in path:
     gameCurrentNode(maze.graph[node[0]][node[1]])
     
-#for i in range(w2):
-#    for j in range(w2):
-#        print(f'Node ({i}, {j}): {maze.graph[i][j].walls}')
-
 running = True
 while running:
     clock.tick(FPS)
